+++ lib/pubsub/rfi_pub_sub.d.ts
@@ -5,12 +5,14 @@ import { GooglePubSub as ApolloPubSubLib } from '@axelspringer/graphql-google-pu
 import { NotificationOf } from '../gql/node-notification';
 import { RfiPubSubConfig } from './pub_sub_config';
 export interface PubEngine extends PubSubEngine {
+    publisher_version: string;
     publishPayload(notificationType: NotificationOf, model: Model, deltas: any[]): void;
 }
 export declare type RfiPubSubEngine = PubEngine & PubSubEngine;
 export declare class RfiPubSub extends ApolloPubSubLib implements RfiPubSubEngine {
     protected topicPrefix: string;
-    constructor(config: RfiPubSubConfig);
+    publisher_version: string;
+    constructor(publisher_version: string, config: RfiPubSubConfig);
     static validatePubSubConfig(config: RfiPubSubConfig): void;
     publish(triggerName: string, payload: any): Promise<void>;
     subscribe(triggerName: string, onMessage: (message: string) => null, options?: Object): Promise<number>;