+++ src/pubsub/rfi_pub_sub.ts
@@ -2,24 +2,31 @@ import { DbModelAndOidScope } from './../db/init-sequelize';
 import { PubSubEngine } from 'type-graphql';
 import { Model } from 'sequelize-typescript';
 
-import { GooglePubSub as ApolloPubSubLib } from '@axelspringer/graphql-google-pubsub';
+import { GooglePubSub } from '@axelspringer/graphql-google-pubsub';
 
 import { publishPayload } from './publishing';
-import { uniqueSubscriptionNamePart } from './helper';
+import { uniqueSubscriptionNamePart, RfiSubscriptionOptions } from './helper';
 import { NotificationOf } from '../gql/node-notification';
 import { status } from '@grpc/grpc-js';
 import { RfiPubSubConfig } from './pub_sub_config';
 import { hostname } from 'os';
 
 export interface PubEngine extends PubSubEngine {
+  publisher_version: string;
   publishPayload(notificationType: NotificationOf, model: Model, deltas: any[]): void;
+  subscribe(
+    triggerName: string,
+    onMessage: (message: string) => null,
+    options?: RfiSubscriptionOptions
+  ): Promise<number>;
 }
 
 export type RfiPubSubEngine = PubEngine & PubSubEngine;
 
-export class RfiPubSub extends ApolloPubSubLib implements RfiPubSubEngine {
+export class RfiPubSub extends GooglePubSub implements RfiPubSubEngine {
   protected topicPrefix: string;
-  constructor(config: RfiPubSubConfig) {
+  public publisher_version: string;
+  constructor(publisher_version: string, config: RfiPubSubConfig) {
     RfiPubSub.validatePubSubConfig(config);
     const { topicPrefix } = config;
     if (config.keyFilename === `/dev/null`) {
@@ -27,6 +34,7 @@ export class RfiPubSub extends ApolloPubSubLib implements RfiPubSubEngine {
     }
     super(config, uniqueSubscriptionNamePart);
     this.topicPrefix = topicPrefix;
+    this.publisher_version = publisher_version;
   }
 
   static validatePubSubConfig(config: RfiPubSubConfig) {
@@ -56,9 +64,7 @@ export class RfiPubSub extends ApolloPubSubLib implements RfiPubSubEngine {
   public async subscribe(
     triggerName: string,
     onMessage: (message: string) => null,
-    // Upstream definition uses Object but tslint does not like that
-    // tslint:disable-next-line: ban-types
-    options?: Object
+    options?: RfiSubscriptionOptions
   ): Promise<number> {
     triggerName = `${this.topicPrefix}_${triggerName}`;
     await this.createTopicIfNotExist(triggerName);