+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file. Starting wi
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+  * publisher_version to the constructor of RfiPubSub
+  * embed the publisher_version to every message that is sent out
+### Removed
+### Changed
+### Fixed
+### Deprecated
+### Security
+
+
 ## [4.0.7] -- 2020-02-19
 
 ### Changed